# RestaurantFinder/src/main/resources/SwagAPIDoc/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Halal Restaurant API","description":"CRUD operations for Halal Restaurants"},{"name":"Prayer Room API","description":"CRUD operations for Prayer Rooms"}],"paths":{"/api/prayer-rooms/{id}":{"get":{"tags":["Prayer Room API"],"summary":"Get a prayer room by ID","operationId":"getPrayerRoomById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successfully retrieved prayer room","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PrayerRoom"}}}},"404":{"description":"Prayer room not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PrayerRoom"}}}}}},"put":{"tags":["Prayer Room API"],"summary":"Update an existing prayer room","operationId":"updatePrayerRoom","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrayerRoom"}}},"required":true},"responses":{"404":{"description":"Prayer room not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PrayerRoom"}}}},"200":{"description":"Successfully updated prayer room","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PrayerRoom"}}}}}},"delete":{"tags":["Prayer Room API"],"summary":"Delete a prayer room","operationId":"deletePrayerRoom","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Prayer room not found"},"204":{"description":"Successfully deleted prayer room"}}}},"/api/halal-restaurants/{id}":{"get":{"tags":["Halal Restaurant API"],"summary":"Get a halal restaurant by ID","operationId":"getRestaurantById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Successfully retrieved restaurant","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HalalRestaurant"}}}},"404":{"description":"Restaurant not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HalalRestaurant"}}}}}},"put":{"tags":["Halal Restaurant API"],"summary":"Update an existing halal restaurant","operationId":"updateRestaurant","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HalalRestaurant"}}},"required":true},"responses":{"200":{"description":"Successfully updated restaurant","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HalalRestaurant"}}}},"404":{"description":"Restaurant not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HalalRestaurant"}}}}}},"delete":{"tags":["Halal Restaurant API"],"summary":"Delete a halal restaurant","operationId":"deleteRestaurant","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Successfully deleted restaurant"},"404":{"description":"Restaurant not found"}}}},"/api/users/register":{"post":{"tags":["user-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/reviews":{"get":{"tags":["review-controller"],"operationId":"getAllReviews","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Review"}}}}}}},"post":{"tags":["review-controller"],"operationId":"addReview","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Review"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Review"}}}}}}},"/api/prayer-rooms":{"get":{"tags":["Prayer Room API"],"summary":"Get all prayer rooms","operationId":"getAllPrayerRooms","responses":{"200":{"description":"Successfully retrieved list","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PrayerRoom"}}}}}}},"post":{"tags":["Prayer Room API"],"summary":"Create a new prayer room","operationId":"createPrayerRoom","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrayerRoom"}}},"required":true},"responses":{"201":{"description":"Successfully created prayer room","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PrayerRoom"}}}}}}},"/api/halal-restaurants":{"get":{"tags":["Halal Restaurant API"],"summary":"Get all halal restaurants","operationId":"getAllRestaurants","responses":{"200":{"description":"Successfully retrieved list","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HalalRestaurant"}}}}}}},"post":{"tags":["Halal Restaurant API"],"summary":"Create a new halal restaurant","operationId":"createRestaurant","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HalalRestaurant"}}},"required":true},"responses":{"201":{"description":"Successfully created restaurant","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HalalRestaurant"}}}}}}},"/api/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}}},"components":{"schemas":{"PrayerRoom":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"location":{"type":"string"},"facilities":{"type":"string"}}},"HalalRestaurant":{"type":"object","properties":{"restaurantId":{"type":"integer","format":"int64"},"name":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"country":{"type":"string"},"rating":{"type":"integer","format":"int32"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"cuisineType":{"type":"string"},"verified":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}},"User":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"passwordHash":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"Review":{"type":"object","properties":{"reviewId":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/User"},"restaurant":{"$ref":"#/components/schemas/HalalRestaurant"},"rating":{"type":"integer","format":"int32"},"reviewText":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}}}}}